JS: Handle git dependenices with filepath sub deps
Some packages are transitioning to yarn workspaces and using local file
paths to refer to sub-dependencies. This breaks npm/yarn git installs so
raising a helpful error.

We could be more clever here and resolve all file paths for git
depencies but seems like this logic should live in npm/yarn if it's
acutally a good idea (not sure atm).