Go (modules): keep bumping pinned dependencies
On second thought, we don't bump indirect dependencies anyway, and if
you're using a top-level dependency I can't see why you'd want to pin it
in the same file using a replace directive, and not receive updates to
it. I suspect most occurences of this are based on misunderstandings
about how Go modules resolution works.

This reverts commit 4d43be5e8cf5f523ac6c46a0c1f2c47ea7008908.