Update dry run to cache editable dependency files
Dumping cached dependency files individually with file extensions making
it possoble to edit the cache and re-run the script. This is helpful
when trying to narrow down a reproduction case.

When running the dry run script it will always dump the dependency files
in: `tmp/account/repo/branch/directory/dependency-files/*`

To read from the cache pass `--cache=files` to the script:
`bin/dry-run.rb npm_and_yarn "dependabot/website" --cache=files`

You can now edit the contents of this folder (make sure it's in sync
with the `cache-manifest.json` file) and re-run against the edited
cache.