Respect HTTP scheme for git dependencies
Even in these modern ages I do have this gem in a git repository that
does not support HTTPS (yet). Currently the uri_with_auth always adds
"https://", which makes the service pack URI invalid. This change makes
sure it uses `http` if and only if the original scheme is HTTP.