Correctly mark requirements as not up to date
Previously when we knew that a newer version was available in the
registry, we didn't correctly mark the dependency as out of date.

`UpdateChecker#requirements_up_to_date?` was first checking if any
dependencies were actually updated, but this check is already happening
in `#can_update_requirements?`, and this was causing us to miss cases
when a new version was _available_ but not possible to update.

Concretely this means that we show a better error when these updates
fail, previously we would mark them as no update needed and now we mark
them as no update possible.