Handle protected branches enforcing linear history
You can enforce linear history for a protected branch preventing any
merge commits to be pushed to the branch, this currently prevents the
pull request from being updated. I'm not entirely sure where we're
creating a merge commit but might be because we're creating a new commit
and force pushing and this is the first rule that gets tripped up in
dotcom.

`BranchProtected` is rescued in dependabot api resulting in a comment on
the pull request explaining why the update failed.