Handle cargo native dependencies
Previously, when specifying a native dependency using `links =
"dependency"`, we would fail as we do not copy over the build.rs file,
which is needed to actually link to the binary. However, to run `cargo
update` we don't actually need to link the dependency, we just need
cargo not to fail. So adding a dummy build.rs file, like we do for
src/main.rs fixes the issue.

Fixes: https://github.com/dependabot/feedback/issues/943