Force actions.yaml syntax check
I expected a CheckSuite to reject an invalid workflow file then it was proposed. AFAICT validation doesn't happen until the workflow file is evaluated (for us: when merged to `main`).

This PR triggers the workflow on _all_ branches, but stops it from doing any actual work.
I'll iterate on the IF syntax here, then bring back the release.