Add vendoring support for go_modules
This builds on the work we've done for go mod tidy, and adds support
for running `go mod vendor` and committing the results.

We can now pass a `vendor` option in the `FileUpdater`, in the bundler
implementation this is not yet used, as we implicitly rely on the
`repo_contents_path` being available there. We should start using that
flag there as well, but leaving it as is for now to make upgrading
easier.