Add filenames and extensions that should not be binary
Noticed a bunch of extensions that were incorrectly being submitted as
binary to the GitHub API.

Alternatively we could also invert this and list a few known binary
extensions, for example:

```ruby
BINARY_EXTS = %w(
  .gem
  .zip
  .tar
  .gz
  .tgz
  .wav
  .rar
  .mp3
)
```

This has the downside that any binary files without an extension are
not picked up. However, we have that same problem currently with text
files. I already spotted some `bin/whatever` files that should not be
submitted as binary to the GH API.