Treat .bundlecache files as binary
The `.bundlecache` file is an empty file that bundler puts in a git
dependencies `vendor/cache/<gem>` folder, to indicate that the file does
not need to be re-downloaded, see:

https://github.com/rubygems/bundler/blob/6394536271a0326efd4ed3544b78e528da7e90b1/lib/bundler/plugin/api/source.rb#L175-L192

Empty files without an extension are treated as binary files by the
`file` util. This is not problematic for submitting to the GitHub API,
since it ends up simply base64 encoding (and subsequently decoding in
GitHub) an empty string.