Ignore go files tagged with `+build`
`go get` will ignore any files marked with `// +build`, so if the
only files present at the root of the directory are tagged with this, we
do not have a valid module, and `go get` will fail.

We now check explicitly if go files are ignored or not, and build a
dummy package if needed.

Co-authored-by: Philip Harrison <feelepxyz@github.com>