[SPIKE] explain why a bundler update was blocked
This returns a message from ForceUpdater that explains why a certain
update was not possible, by traversing the dependency tree and finding
all parent dependencies that are not satisfied by the target version.

Few things to improve:
- The message is now formatted in the native helper, I would prefer to
return raw data from there, and leave any formatting to whatever drives
dependabot.
- Because we currently depend on the native helper returning an error,
it is hard to pass data from the native function to dependabot-core.
- We need to call `latest_version_resolvable_with_full_unlock?` in order
to hit the error that sets this new message. I'd prefer to be able to
just call a method on the UpdateChecker instead.
- I dislike having to string compare against the native errors `error_type`, it
would be nice to have more granular errors that we can return from
there.