Fix dependabot-core-ci build by removing buildkit
Running docker build for the dependabot-core-ci image was hanging at the
"exporting layers" step after running all the steps in the Dockerfile.

This seems to indicate that docker buildkit is doing a bunch of heavy
lifting/diffing/hashing of the content (lots of bundle files) which is
too memory intensive and causes the docker build command to exit 137
(probably OOM).