Only tally installed/updated/uninstalled repo's in InstallationManager
In our custom DependabotInstallationManager we don't care for any of the
work that the normal installation manager does, we just want to see
which packages can be updated/removed etc. This change guts out most of
the work in `execute` so that we only call our custom handlers, which no
longer call their parent.

This does introduce a test failure locally, so will need some work.