Fix JS debugging in vscode
This adds a couple launch scripts to debug npm and yarn tests when
running then using jest.

The launch scripts complained about a missing config file so added this
and changed the default env from jsdom to node as this is what we use.

Changed `eslint-plugin-prettier` to `eslint-config-prettier` as this
seems like the recommended extension now when using prettier with
eslint: https://prettier.io/docs/en/integrating-with-linters.html

This extracts these changes from the npm 7 branch:
https://github.com/dependabot/dependabot-core/pull/3011