Npm7: correctly parse version from package dependencies
Previously, when using npm 7 workspaces, when a dependency was installed
from a workspaces' `package.lock` but a _different_ version of that
dependency was installed via the top-level `package.json`, we would grab
the top-levels version, resulting in incorrect behavior.

This changes the implementation to look in the lockfiles' `packages`
object for the actual dependency from that workspace, and grab the we
need there.