npm: Use CLI for peer dep conflicts, and default to it without lockfile
TLDR: Replace the native peer dependency conflict helper with the npm 7
CLI, and use that version when there is no lockfile.

I've been investigating a slow update scenario for what seems like a
fairly normal scenario. A project with 14 dependencies[1], and no
lockfile present.

From the logs I noticed a ton of requests being made to yarnpkg.com, and
at first I couldn't really place those.

Although there seems nothing weird about this setup, running an update on my
local machine is pretty slow: `7m2.660s`

My laptop is a lot beefier than the average dependabot worker, so it seems like
this update is indeed slow. It's not clear to me at all why that's the case
though, it's a completely normal setup. None of these packages even seem to
need an update, so taking 7 minutes to tell us that on a beefy laptop seems
excessive.

To figure out what's going on, I ran [stackprof](https://github.com/tmm1/stackprof)
on the dryrun for a single update (`axios`) and noticed that a _lot_ of time is
spent in `open3`, which runs our native helpers:

<img width="1014" alt="Screenshot 2021-02-11 at 22 26 05" src="https://user-images.githubusercontent.com/749864/107700907-71905a00-6cb8-11eb-8567-a2959518624b.png">

From tracing the code I noticed that we're running the yarn specific
`yarn:checkPeerDependencies` when we don't have a lockfile. That explains all
the requests to yarnpkg.com. I didn't quite understand why we use the yarn
resolver and it seemed pretty slow, I tried replacing it with the npm 6
native helper, but it misses some edgecases when there is no lockfile. I
then tried to replace it with the npm 7 native helper, which runs
Arborist, but this was roughly as slow.

I wondered why this is so slow, and ran some comparisons against using
the (npm7) CLI to just perform an `npm install`, and this seemed a bit
faster and always correct.

This PR replaces the npm 7 native helper with a call to the CLI, and it
also defaults to using this version when there is no lockfile present.

This seems to consistently shave off ~20 seconds from a 54 second update for a single dependency with no lockfile:

Before:

```
real	0m54.637s
user	0m29.511s
sys	0m19.109s
```

<img width="976" alt="Screenshot 2021-02-16 at 14 22 25" src="https://user-images.githubusercontent.com/749864/108068495-752e3300-7062-11eb-9180-118fe219d3d1.png">

After:

```
real	0m35.279s
user	0m19.014s
sys	0m5.420s
```

<img width="973" alt="Screenshot 2021-02-16 at 14 22 38" src="https://user-images.githubusercontent.com/749864/108068552-85461280-7062-11eb-95bb-224b54ba78a2.png">

[1]: `axios, change-case, dayjs, jstz, logrocket, react, react-dom,
react-ga, react-scripts, sass, title-case, babel-eslint,
eslint-config-prettier, prettier`