Bundler v2 [pre-release]: Add and test jfrog source helper
Adds a basic test for the jfrog_source method. I tried to test this from
the core codebase where it's used as well, but due to the current setup
it's currently very hard to do that without connecting to an actual
jfrog server. Since we had no coverage for this before, I'm calling it
an improvement.