Terraform: Fix updating multiple providers
Previously, the regex we used to figure out which part of the codebase
to run the replace on was too broad, and would include the entire
`required_providers` block, which results in the replace only working on
the first provider in the file.

This adds a new regex to use for provider updates, which will find the
specific provider within the `required_providers` block.

Fixes https://github.com/dependabot/dependabot-core/issues/3760