Terraform: handle implicit (v0.12 style) provider sources
When leaving out a `source = ..` declaration for a Terraform provider,
it falls back to assuming the default `hashicorp` namespace and the
declared local name as the package name.

This is unrecommended, but we do come across it.

Previously Dependabot would fail, with this change it parses + updates
these cases correctly.

Co-authored-by: Mo Khan <xlgmokha@github.com>