Yarn: use .yarnrc file if present
Fixes a bug when trying to force yarn to use `registry.npmjs.org`
instead of `registry.yarnpkg.com`.

Previously if a .yarnrc file was present with registry `registry
"https://registry.npmjs.org"` it would be ignored during updates as we
run yarn programtically and didn't instruct it to use the config file.

I've also tested a dry-run without a `.yarnrc` file present.