Tests: avoid squatted repositories
A few repositories we expect not to exist on npm have been squatted.
Where possible this updates those tests to use a namespace that we own,
and for one fixture we explicitly tested against an unscoped package,
those tests have been removed as we do not expect behavior to differ
between scoped and unscoped packages.