Terraform: fix module updates with a lockfile
This is the other side of https://github.com/dependabot/dependabot-core/pull/3968

This fixes module updates when there's a lockfile present. Modules are
not included in the lockfile so adding a guard to only attempt lockfile
updates for providers.

Fixed some test setup to match what the parsed files would look like.