npm: prevent scoped registries becoming global
Fixes a bug where all packages (public and private) are forced through a
private registry, where it's only intended to be used for scoped
packages so fails for public packages hosted on npm.

Added a check to the method that attempts to extract a global registry,
checking if credentials registry host has already been configured in
`.npmrc` as a scoped registry.

This prevents dependabot from re-writing the `.npmrc` global `registry`
with the scoped registry:

```
registry=registry.npmjs.org
@private=private.example.com
registry=private.example.com
auth..
```