Pass full version to bundler as `BUNDLER_VERSION`
In a recent version of bundler, passing a major version like
`BUNDLER_VERSION=2 bundle --version` no longer works.

I suspect this may have happened here: https://github.com/rubygems/rubygems/pull/4750/files

We now need to pass the exact version of bundler we're using, which is a
bit cumbersome as you can see from the diff, and it also means we'll
have to change a few more lines every time we update a version of
bundler.

I don't think the current behavior we were relying on was documented, so
I'm not sure if we can open an issue with the Bundler team.

This seems like the most straight forward fix to me, but I am definitely
open to other suggestions