Maven: Handle nested plugin declarations
This was originally resolved in
https://github.com/dependabot/dependabot-core/pull/4417 but later
reverted, because that changed introduced a regression due to parsing
the XML using Nokogiri, which caused formatting of the XML to be lost.

We've observed errors around pom files that have a nested plugin
section inside an outer plugin declaration.

Given the following XML:

```xml
<plugins>
  <plugin>
    <configuration>
      <jvmTarget>11</jvmTarget>
      <compilerPlugins>
        <plugin>spring</plugin>
      </compilerPlugins>
    </configuration>
    <groupId>org.jetbrains.kotlin</groupId>
    <artifactId>kotlin-maven-plugin</artifactId>
    <version>${kotlin.version}</version>
  </plugin>
</plugin>
```

The <plugin>spring</plugin> declaration would cause a regex match,
resulting in the required information (the version in this case) to be
omitted from the XML snippet.

This is resolved by matching on an additional nested plugin block in the
regex, so that we always end up matching until the next closing tag.

This does mean that if _another_ level of nesting is present we'd still
end up with an incomplete XML section. This also does not resolve the
issue for other elements than the `plugin` one. If we do observe the
same issue with those other elements we can introduce the same type of
nesting there.