For testing PoetryFileUpdater, use .prepared_pyproject
...rather than testing against the updated files as in the previous commit.
With this approach, we don't need to set up an actual repository with credentials,
since all we are doing is checking that the credentials do not leak into the
pyproject file.

Since the lock file is created by Poetry itself from the pyproject file,
if our pyproject file is clean of credentials when we generate the lock file,
the only way credentials could then leak in would be a bug in Poetry.