Docker: Ignore ARGs without default values
In https://github.com/dependabot/dependabot-core/pull/4598 we added
support for ARG FROM. However, this assumed that a default value is
always specified, when it isn't the code would crash.

This is not that common, however:

Due to how we parse Dockerfiles, this actually also happened for regular
`ARG`s in a Dockerfile, even if it didn't belong to the `FROM`, meaning
that any ARG without a default would cause our update to fail.