Pub: use system root certificates
By default Dart uses it's own certificate store. This is problematic for
us in production where we send traffic through a proxy to add
authentication headers to requests being made, which uses a randomly
generated certificate that we add to the systems cert store, which will
then be ignored by dart.

This instructs dart to use the system certificates instead.