Remove NOTE not totally accurate
This note makes total sense for `helpers/v1/build`, because there we
explicitly pass `BUNDLER_VERSION=1.17.3` to `bundle install`, which
generates a `Gemfile.lock` using that version, which indeed forces
RubyGems into choosing Bundler 1.17.3 instead of the highest version
installed.

Here, however, we would get the same behaviour of 2.3.18 being chosen,
with or without a `Gemfile.lock` using that, because it's the highest
version.