Move comment to relevant section of code
Previously, it was easy to lose track of this comment because it's
located quite a ways from the code. Moving it makes it less likely that
packages will listed in the wrong category.

Although I'd happily bet an ice cream cone that the comment is
partially incorrect and some of the listed packages are not just for
python.