Make Dependabot watch the per-ecosystem gemspecs
Previously Dependabot watched the top-level gemspec, but not the
per-ecosystem gemspecs.

Historically we haven't had much (any?) per-ecosystem gem requirements,
so this hasn't been a problem before.

But it's probably wiser if we start watching these. Especially because
for the Ruby 3.1 upgrade the `pub` ecosystem will likely need to add a
specific dependency on `webrick`.