Add a `.ruby-version` file
We're starting to pin the desired ruby version in multiple places:
* Rubocop
* the new `ruby/setup-ruby` action that we're adding in https://github.com/dependabot/dependabot-core/pull/5433

We used to not ship the `.ruby-version` file, with the thought that this
is a library... but this file isn't included in the gemfile spec, so
we're not tying the hands of our users. This merely sets the default
version of Ruby that we're using for development. We will keep this in
sync with the version of Ruby installed in the Dockerfile, so it'll
target the version of Ruby we use in production at GitHub. But since not
included in the gem, others are free to use a different Ruby version.

Copy the `.ruby-version` file into the CI container.