Always drop the fake go file
As @jakecoffman pointed out in
https://github.com/dependabot/dependabot-core/pull/4954#issuecomment-1090283390,
checking for the presence of go build tags is relatively brittle and the
existing check isn't thorough enough.

So instead always write the temp file to ensure `go get` always
succeeds.

This both hardens and simplifies the code logic, but means that instead
of rarely writing/deleting the temp file, we now always write/delete the
temp file.