Decouple Bundler versions
On the one hand, the version of Bundler that runs dependabot itself.
This one can be really anything really, but it's (or may be) restricted
by manifest files (`Gemfile` and `Gemfile.lock`) files in the
dependabot-updater repository.

Then there's the version of Bundler that runs Bundler native helpers.
This one is restricted by the version of Bundler in the manifest files
being updated.

With the introduction of RubyGems 3.3, the distinction between these
will be more messy because some Bundler behavior to automatically switch
to our more appropriate versions is activated.

This commit tries to decouple these two. Native helpers will have their
own home, with the Bundler version each of them need, while dependabot
itself will use the Bundler version installed globally.