Ignore newlines when determining indentation
Recently it was reported that Dependabot had changed the indentation for
an entire package-lock.json from two spaces to one. Upon investigation,
it was determined that a newline had been added to the end of that file
(by the user), and Dependabot had considered the newline a single space.

By using `[[:blank:]]`, we get the benefits of `\s`, but with a little
less aggressiveness.