Merge previously private updater code into core
This adds a new component to dependabot-core that is only used and
useful to GitHub internally, it's part of how we run Dependabot in our
internal systems. By merging this into the main codebase we hope to be
able to simplify both it and how core works going forward, and it will
open up some simplifications in our deployment and development processes.

This code is not useful to anyone outside of GitHub, other than maybe
informative in how to run dependabot-core in a service setting. It's
also explicitly _not_ a goal for this code to be useful or reusable
outside of its current context.

One thing that _is_ useful to outside contributors, is that this change
also includes a set of smoke tests that run against both the core
ecosystems and the updater. This will give us higher confidence when
reviewing + merging PRs.

The code in this PR was contributed to by many people, and it has a rich
history available in an internal repository. Since those commits were
made in a private context, we've squashed all of them in a single
commit and omitted any non GitHub-owned email-addresses.

Co-authored-by: Barry Gordon <896971+brrygrdn@users.noreply.github.com>
Co-authored-by: David McIntosh <804610+mctofu@users.noreply.github.com>
Co-authored-by: Grey Baker <greysteil@github.com>
Co-authored-by: Harry Marr <hmarr@github.com>
Co-authored-by: Jake Coffman <jakecoffman@github.com>
Co-authored-by: Jason Rudolph <jasonrudolph@github.com>
Co-authored-by: Jeff Widman <jeffwidman@github.com>
Co-authored-by: Jurre Stender <jurre@github.com>
Co-authored-by: Landon Grindheim <landongrindheim@github.com>
Co-authored-by: Lane Seppala <lseppala@github.com>
Co-authored-by: Mattt <mattt@github.com>
Co-authored-by: Nickolas Means <nmeans@github.com>
Co-authored-by: Nish Sinha <nishnha@github.com>
Co-authored-by: Pete Wagner <1559510+thepwagner@users.noreply.github.com>
Co-authored-by: Philip Harrison <feelepxyz@github.com>
Co-authored-by: mo khan <xlgmokha@github.com>