Add new interface to register experiments
Currently experiments are required to be passed in as an optional
`options` flag to the individual classes like the FileFetcher and
FileUpdater. This is fine, but it becomes a bit cumbersome when more
deeply nested objects need to access this flag, as we need to pass it
down to all of them.

This introduces a global `Dependabot::Experiments` interface that we can
use to register experiments and access them where needed.

The idea is that these are set in the Updater before running an update,
and will be available for the duration of the job.