Preserve requirement source when updating transtive dep parents
We had been setting the source of the parents to nil but this caused the
FileUpdater to think the requirement was always changing. This might not
be the case for a library with a requirement range that already allows
the fixed version. We should prefer to preserve the existing source when
an updated source is not provided.