Fix crashes on Python libraries using multiple manifests
If a library keeps a requirement duplicated in
pyproject.toml/poetry.lock and a requirements.txt file, Dependabot
updates will crash because we will try to widen requirements in the
requirements file, which is not supported.

Honestly, I don't know what users of this kind of setup expect so I
opted for fixing the crash without changing the current behavior,
namely, no updates.