Remove YarnLockfileParser
YarnLockfileParser is a vestige from an era in which it may have been safer to parse a yarn.lock ourselves rather than shelling out to a native helper. Elsewhere Dependabot uses a native helper, `yarn:parseLockfile` (which in turn uses Yarn's own lockfile parser) to parse yarn lockfiles. Since the YarnLockfileParser is no longer required, and in order to parse lockfiles consistently, the class is being removed.