Move location of `hex` version pin
Moves the location of the `hex` version pin to a more intuitive
location.

This also centralizes the install logic within the helper builder
script so that we don't call `mix local.hex` twice.

We could do it externally, but Hex requires being installed as the
Dependabot user, and so we end up with the Hex install commands sitting
right next to the build script... so rather than scattering the logic
across the Dockerfile and build script, I opted to keep it all in the
build script. This way if a developer is using the build script for
local development w/o rebuilding the docker image, it's a repeatable
experience.