Remove CODEOWNERS
This file is actually incorrect, since it only requires reviews when
first level files in each folder are changed.

This is easy to fix, we just need to remove all asterisks. See
[examples].

However, I think fixing the file in the current situation is going to
get a bit annoying, because a lot of review request are going to start
being request to teams without any members.

So remove all the sub-teams from the file for now.

[examples]: https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#example-of-a-codeowners-file