Fix dependency parsing when a gemspec loads another
If a gemspec loads dependencies from another gemspec dynamically, then
those dependencies would have requirements for both gemspec, the one
that really defines it, and the one that loads it.

This is currently the case for our own gemspec files in dependabot-core.

When checking whether the requirements of a dependency can be unlocked,
Dependabot checks whether modifying all requirements of the dependency
always results in a changed manifest. However, in the case of these
dynamically loaded gemspecs, this replacement does not change the
manifest, and as a consequence, Dependabot interprets that the
requirement cannot be unlocked and is unable to update the dependency.

As a solution, when deciding whether to add gemspec requirements to a
dependency, we parse the gemspec and look for explicitly calls to
`add_dependency` or other gemspec DSL methods with the name of the
dependency as an argument.