Fix branch with a "version-like" name confusing Dependabot
Current `actions/checkout` repo contains a branch named
`dependabot/npm_and_yarn/minimatch-3.1.2`. Since a while ago, Dependabot
considers branches for resolution, because it's a fairly popular action
release management style to use versioned branches. Also, the git commit
checker supports version branches with custom prefixes, thus making
something like `dependabot/npm_and_yarn/minimatch-3.1.2` be considered
as a candidate. In addition to all that, the version associated with it
is higher than the highest version of the action, so 3.1.2 mistakenly
wins as the highest version.

However, since 3.1.2 does not really exist as a release, Dependabot
falls back to updating to the latest commit in the branch the existing
commit belongs to. But since in this case, it's already pointing to a
tag, it leaves it untouched.

The problem here is that branches should not be considered at all when
we are pinned to a tagged commit SHA. In this particular case, we only
want to upgrade to other pinned commit SHAs.

After this change, we get the expected result again:

```
=== actions/checkout (1.1.0)
 => checking for updates 1/1
🌍 https://github.com/actions/checkout.git/info/refs
🌍 https://github.com/actions/checkout.git/info/refs
 => latest available version is 3.1.0
 => latest allowed version is 3.1.0
 => requirements to unlock: own
 => requirements update strategy:
 => updating actions/checkout from 1.1.0 to 3.1.0

    ± .github/workflows/howie-on-issues.yml
    ~~~
    30c30
    <       - uses: actions/checkout@0b496e91ec7ae4428c3ed2eeb4c3a40df431f2cc
    ---
    >       - uses: actions/checkout@93ea575cb5d8a053eaa0ac8fa3b40d7e05a33cc8
    ~~~
🌍 Total requests made: '5'
```