Ignore URLs in Actions
Dependabot currently supports version updates for Actions that are referenced as a GitHub repository.
There is a feature request to support image/container URLs in GitHub Actions, which will require checking which
versions of the Action are available on the registry.

In the mean time, Dependabot should stop trying to parse these URLs as a reference to a GitHub repo and ignore them.