Update `monolog/monolog` fixture to latest packagist response
For a few upcoming PR's, it'll be helpful if the `monolog/monolog`
packagist fixture is current to match what's _actually_ returned by
https://repo.packagist.org/p/monolog/monolog.json.

Initially I updated all the other fixtures as well, but the
`illuminate/*` fixtures ballooned to 43+ MB... and then when I switched
them to the `v2` metadata endpoint (https://github.com/dependabot/dependabot-core/issues/3010) they shrank
back down... no point in adding that much cruft to our git history when
I'm going to almost immediately drop it.

But the `monolog/monolog` change is necessary because I'll be updating
all the tests that point at the generic `packagist_response.json` to
point at this fixture... and since it already has a newer version than
is in `packagist_response.json`, might as well first ensure the fixture
is fully-up-to-date.