Clarify `composer` `v1` `PEAR` unit test
It took me a while to understand this unit test, because this package
_is_ actually present on packagist: https://repo.packagist.org/p/pear-pear.horde.org/horde_date.json

However, from the associated `composer.json` fixture plus the existing
fixture, it's clear that this test is checking what happens when a PEAR
dependency is _not_ present on packagist.

So remove the fixture from the `packagist_responses` directory which is
supposed to match what Packagist actually returns and hardcode the
response for this single unit test to avoid confusion of future devs
thinking they need to update it to match the actual packagist response.

This will make the https://github.com/dependabot/dependabot-core/pull/6315
a bit more straightforward because composer v1 doesn't know about the v2
metadata API so I will need to hardcode both v1 and v2 metadata
responses (which return different HTTP status codes).

PEAR is mostly dead it seems, but it doesn't hurt to leave it around in
case any old composer v1 projects still have PEAR deps listed... and
this will all be gone soon enough once we remove support for `composer`
`v1`:
* https://github.com/dependabot/dependabot-core/issues/6298