only replace url for private pipenv source (#6158)
Instead of filtering out private pipenv sources completely, and
inserting the urls from the Dependabot credentials configuration (which
removes other parts of the pipenv source configuration, e.g. the source
name), only replace the url with the new authed url when possible.
Also adds the credentials to the Pipfile as-is.