Added support for Azure in git_submodules (#6320)
Support `git_submodules` from Azure DevOps. All the code was present but not set up.

Changes in #6304 and #6306 helped a great deal too in handling submodule URLs for Azure DevOps:

```bash
[submodule "another-repo"]
	path = another-repo
	url = https://contoso@dev.azure.com/contoso/MyProject/_git/another-repo
```

The suffix `.git` does not get added in Azure DevOps. For GitHub / GitLab / BitBucket, it works with and without the suffix. Originally I explored removing the suffix entirely, but it's a larger change, so I opted to exclude it only for `dev.azure.com` and `*.visualstudio.com` hostnames as there is no other way to tell if we should. This will not work for Azure DevOps Server but maybe that is a fight for another day.

Fixes: #6340 
Fixes: #5756