Fix updating yarn.lock on yarn berry projects with workspaces
If a dependency has a bad checksum, we fail to `yarn install
--mode=skip-build` even when yarn cache is not necessary because of not
being on any situation needing the cache.

We fix that by explicitly configuring a global cache the first time we
detect we won't need it because otherwise yarn runs may generate a cache
and result in further commands running in `--mode=skip-build`.