remove bundler cache after installing (#6662)
We're finding that the Bundler ecosystem is evading the smoke tests after a while. This is probably due to how Bundler caches dependency data. When we run `bundle install` during a build, it stores a cache of the versions it fetched under `~/.bundle`. After a day or so this is out of date and Bundler attempts to get new data, probably sending new headers to do so, breaking the cache.

So by removing this cache, our tests should always make the same call to get fresh data.