Adjust poetry python version to match installed in version resolver
Resolves this error when checking for updates when a specific python
version is pinned in pyproject.toml.

Dependabot::SharedHelpers::HelperSubprocessFailed:
The currently activated Python version 3.11.2 is not supported by the project (3.11.1).
Trying to find and use a compatible version.

Poetry was unable to find a compatible version. If you have one, you can explicitly use it via the "env use" command.
 ./lib/dependabot/python/update_checker/poetry_version_resolver.rb:349:in `run_poetry_command'
 ./lib/dependabot/python/update_checker/poetry_version_resolver.rb:167:in `run_poetry_update_command'