Rewrite README to be more n00b friendly (#6955)
Over time the Readme has grown organically by adding new sections as
needed. Unfortunately, the length and relative lack of order makes it
not very approachable for folks new to the project.

So this adds a Table of Contents structured in a way that's more
accessible, and then re-orders the content appropriately.