Use `skip()` instead of `pending()` where appropriate (#6779)
In `rspec`, `skip()` completely skips the test. Whereas `pending()`
runs the test, and expects it to fail. In fact, if it doesn't fail,
`pending()` will fail the test suite so that the developer realizes
whatever the failure mode was is now fixed.

For tests that we _never_ expect to pass without developer time, there's
no point in marking them as `pending()`, we should mark them as `skip()`
instead.

I grep'd for `pending(` and switched to `skip(` where appropriate.