Fix incorrect detection of top level gemspecs
When dealing with libraries, Dependabot sets an extra requirement on Ruby to
be the lowest version supported by the gem.

However, it was also incorrectly doing so when dealing with vendored
path gems, because all gemspecs were being considered when adding this
extra requirement, not just the one in the top level.