Add tag to indicate if repo is private in sentry reports
Currently finding a way to reproduce issues found in Sentry can be
challenging, usually the easiest way is to find a public project that we
can run our code against. Today this is hard, because we need to
manually comb the errors in sentry until we find one that's public.

With this change, that should no longer be an issue.