Fix outdated reference to DockerHub (#7271)
When the images were split, this was updated to push to GHCR rather than Dockerhub.

In fact, you can see here the last image was pushed 3 months ago: https://hub.docker.com/r/dependabot/dependabot-core/tags

But this reference to dockerhub wasn't updated.

Note:
This pushes the `dependabot-updater-core` image to GHCR, but not the ecosystem-specific images... that was likely an oversight. Correcting that is outside the scope of this PR though.