Fix edge case when updating Actions with mixed versions (#7410)
Previously we would avoid updating requirements if the commit for the
dependency does not match the commit for the updated reference.

The commit for the dependency is guessed by the `GitCommitChecker` using
the _first_ source as the dependency source. However, this can be
inaccurate when updating dependencies with sources with mixed versions,
since they may point to different commits.

I think this is an issue in the `GitCommitChecker` but I also noticed
that these checks should not be necessary, we can unconditionally update
the source, even if the commits don't match.