Workaround `pyyaml` / `cython` incompatibility
This test was failing when it tried to install `pyyaml`:
```
rspec ./spec/dependabot/python/update_checker/pip_compile_version_resolver_spec.rb:139
```

The underlying issue is `PyYAML` isn't yet compatible with `cython` `3.0.0`.
So manually I bumped the `pyyaml` pin to `6.0.1` which includes a
holdback pin to not install the latest `cython`:
* https://github.com/yaml/pyyaml/commit/c42fa3bff1eabdb64763bb1526d9ea1ccb708479

The better way to fix this would be to regenerate the `pip-compile`
output lockfile, but I wasn't sure I'd be able to do that since the
`imports_shared.in` file includes a recursive reference to another
requirements file... and that other requirements file doesn't appear to
exist on disk in the fixtures... from a quick glance at the test specs
code, it may dynamically generate that file. So I'd have to start it,
then drop into a debugger to pause it, then manually go look at the file
(if it's even saved to disk), and then re-run it... and even then, it
may cause other failures with the test since it updates all the other
versions...

Anyway, rather than tracking all that down, I merely tried the pin, and
low and behold it fixed the test. So I think we're good for now,
especially as `pyyaml` isn't part of the test at all, it's just a
transitive dep of the actual package under test.