Use `msbuild /pp` to preprocess project files
The MSBuild preprocess option will give us the definitive list of `.props`/`.targets` files that a project imports so we're guaranteed to only look at the relevant files.

This necessitated the revert of a previous PR to use XPath to check some project files, because the MSBuild tool errored out on invalid MSBuild XML.  At least as far back as 2008 this XML was invalid, so it's not a case of us needing to support older project files, but the file in question is entirely invalid and _should_ result in an error.