Parse repositories from the top-level buildfile
Gradle was previously ignoring repositories from the top level buildfile (build.gradle.kts) which resulted in Dependabot looking for the dependency in the wrong repository

Some repositories are still not getting picked up correctly, specifically ones that use a custom URI like https://kotlin.bintray.com/kotlin-dev/

The bintray repository has been sunset, so we should eventually remove it. The jCenter repository is marked as read-only currently and will likely be sunset in the future.