Request gzip'ed responses from rubygems API
I noticed that we're spending a lot of time making requests to rubygems,
fair game, but I also noticed that we're requesting uncompressed JSON
blobs, and JSON tends to compress relatively well. I'm seeing a modest
performance improvement in local testing on a small-ish repo.
Benchmarking network requests is always a little tricky, but I think
it's fair to assume this certainly wouldn't hurt performance.

Excon is kind enough to handle the decompression for us.