Enable YJIT for even-numbered update jobs
We primarily upgraded to Ruby 3.3 so that we could enable YJIT. We did
this in order to hopefully mitigate some of the performance issues some
of our larger customers have faced.

This change enables YJIT for _some_ update jobs. The reason for the
incremental enablement is that we would like to measure the impact that
YJIT has before adopting it fully (which we intend to do).