Add `#directory` to Dependabot::Dependency
This is a concept which has been deemed to be missing. The effect of
this is that when a pull request exists for a group, without there being
knowledge of a directory, individual pull requests are then created. By
adding `#directory` to Dependabot::Dependency, we'll be able to prevent
this.